+++ src/decorators.ts
@@ -34,7 +34,7 @@ import {
   $reusableCorrelationId,
   $server,
 } from "./namespace-decorators.js";
-import { $jsonSchemaExtension } from "./extension-decorators.js";
+import { $extension, $jsonSchemaExtension } from "./extension-decorators.js";
 import {
   $useBinding,
   $useChannelBinding,
@@ -59,6 +59,7 @@ export const $decorators = {
     operationId: $operationId,
     operationSecurity: $operationSecurity,
     operationTrait: $operationTrait,
+    extension: $extension,
     jsonSchemaExtension: $jsonSchemaExtension,
     parameter: $parameter,
     protocol: $protocol,